Add support for layering_check
This feature currently applies to C/C++ compiles. It validates that for
every header you import you have a corresponding target in your deps
that surfaces that header. Bazel handles a lot of this logic internally
in the CC rules but it's up to us to provide a system-level modulemap
file that contains all system headers, so that non-system headers that
aren't in your deps are correctly flagged as a layering check violation.

To do this we use find to discover all headers in the current
DEVELOPER_DIR. There are a few potential caching issues surfaced with
this:

1. If the paths were absolute, the Xcode path would be an input. We
   avoid this by storing relative paths, and then using a vfsoverlay to
   make it "look" like the modulemap is inside of Xcode itself, which is
   the only way to make relative paths work in modulemaps.
2. If different Xcode versions have different headers, and you support
   multiple Xcode versions in your build, the modulemap file is an input
   and will cause caching issues. I think the future solution to this
   would be allow users to provide their own modulemap with a constant
   set of headers, since you likely won't care about the new or removed
   ones (and if you do you have to only support 1 of the Xcode verisons)

This is currently disabled by default and can be enabled with
`--repo_env=APPLE_SUPPORT_LAYERING_CHECK_BETA=1`. This is primarily
because of the issues above, and that generating the modulemap file
takes ~5 seconds.